+++ contracts/ISwapRouter.sol
@@ -2,14 +2,14 @@
 
 pragma solidity 0.5.17;
 
-
 interface SwapRouter {
     function WETH() external pure returns (address);
+
     function swapExactTokensForTokens(
-      uint amountIn,
-      uint amountOutMin,
-      address[] calldata path,
-      address to,
-      uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 amountIn,
+        uint256 amountOutMin,
+        address[] calldata path,
+        address to,
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 }
